another refactor of builder
